Fix on_pull_request GitHub flow helper method (#144)
