+++ libraries/git/steps/git_distributions.groovy
@@ -68,5 +68,29 @@ void init_env(){
 }
 
 def fetch(){
-  return getStep(env.GIT_LIBRARY_DISTRUBITION)
+  def stepName = env.GIT_LIBRARY_DISTRUBITION
+
+  if(jteVersion.lessThanOrEqualTo("2.0.4")){
+    return getBinding().getStep(stepName)
+  } else {
+    return this.getStepFromCollector(stepName)
+  }
+}
+
+@NonCPS
+def getStepFromCollector(String stepName){
+  try {
+    Class collector = Class.forName("org.boozallen.plugins.jte.init.primitives.TemplatePrimitiveCollector")
+    List steps = collector.current().getSteps(stepName)
+
+    if(steps.size() == 0){
+      error "Step '${stepName}' not found."
+    } else if (steps.size() > 1){
+      error "Ambiguous step name '${stepName}'. Found multiple steps by that name."
+    } else {
+      return steps.first().getValue(null)
+    }
+  } catch(ClassNotFoundException ex) {
+    error "can't find the TemplatePrimitiveCollector class. That's odd. current JTE version is '${jteVersion.get()}'. You should submit an issue at https://github.com/jenkinsci/templating-engine-plugin/issues/new/choose"
+  }
 }
